refactor(mindweaver): relocate proto generation to consumer directory
Move generated protobuf files from packages/pkg/gen/proto to
packages/mindweaver/gen/proto to colocate with the sole consumer.

This change:
- Generates protos directly to mindweaver package
- Removes shared pkg/gen location (no actual sharing occurred)
- Enables fail-fast on breaking API changes at build time
- Follows pattern already established by neoweaver client
- Updates buf.gen.yaml to target new output location

Generated files remain committed for developer convenience.
CI will verify freshness in future step.
